Committer interface
